e17 now respects default application settings in file manager.
load defaults.list from ~/.local/share/applications and
$XDG_DATA_DIRS/share/applications.

This file is a FreeDesktop.Org standard used by both Gnome and KDE. It
is already managed by E17 at Settings > Apps > Default Applications.

The standard allows a list of applications to be specified for each
mime-type. This order is now used by e_fwin.

If there was a list specified then "Open" will use the highest
priority instead of asking which one to use. "Open with..." still
works and respect the order.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@73435 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33