enlightenment welcomes elementary.
Now it's possible to do elm_win_add() from your e_modules and use all
elementary features instead of painful e_widgets.

We should consider this experimental, but after e17 is out we should
start to convert dialogs and such to elementary, eventually
deprecating e_widgets and the theme duplication.

NOTE: requires r74156 with elementary patch.

I'll send an email to the list with a simple patch to skel module to
demo this.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@74157 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33