naviframe: improve default back button.
back button now uses standard icon "arrow_left", and provides a "Back"
text that should be translated with the application.

Then we can define the elm/button/base/naviframe/back_btn/default as
an alias to the default button. If a theme wants it could still
provide an alternative button. IOW no breaks.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@74399 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33