naviframe improvements:
 - fixed theme to look less like an arse:
    - overlap theme is translucent
    - naviframe title visibility toggle is now animated
    - item pop/push animation now applies to title area as well
    - icon without title label is now centered (see Naviframe 2 test)
    - title label and subtitle are now centered among the free space, should not overlap buttons anymore.
    - title will use a font-size range, trying to not overlap buttons anymore.
    - title will use ellipsis, trying to not overlap buttons anymore.
    - prev/next buttons do not have a fixed min/max size anymore
 - emit signals elm,state,title_label,show and elm,state,title_label,hide
 - by default the previous button label is the title of the previous page

however by using a regular elm_button with an icon and long label
exposes a bug with button theme. And there is a conceptual issue: if
the previous button label is too big, it will look horrible. We should
have a maximum size that, if reached, should have ellipsis. This
should be a different elm_button style, with
elm/button/base/naviframe/back_btn/default being an alias to it.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@74403 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33