elementary: add thaw/freeze to Elm_Layout and use it in Elm_Calendar.
NOTE: This improve calendar performance quite a lot, but a better improvement
would be to populate the object when we really need it. I don't know how to
do that in current Elementary source code. Reason for another expected massive
improvement is that every time you set a property on an Elm_Calendar trigger
a populate and that one is costly (Around 6 populates call per Elm_Calendar in
our code example, meaning doing it could give us a 6 times boost...).


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@74553 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33