From: Kim Shinwoo <kimcinoo.efl@gmail.com> Subject: [E-devel] [patch][elementary] dayselector - focus chain, and about accessibility policy.
dayselector does not work in focus cycle because there is no
focus_chain().
because dayselctor is a child of Elm_Layout_Smart_Class, so it is
enough to
use focus_chain() of layout.

and.. very carefully..
i'd like to raise an issue that currently visually impaired person
should
be obliged to click(tap) quickly two times to activate widget such as
button, check.
the implementation to check 2nd click in elm_access is great. but one
single click is enough to activate button or change state of elm_check.
moreover some widgets like dayselector which has customized style for
elm_check does not work properly, because they do something at signal
'mouse,down,1'.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@74649 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33