Elm: Be sure to set a size for the window.
Previously, if you clicked the icon animation test if would appear to
fail as the window was being created with a 0 size (and never
resized). This just makes sure that the window gets resized to
"something" so that it will actually show up (rather than a user just
blindly clicking the test and wondering why nothing is happening).



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@75525 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33