Elm: Fix slider popup to account for framespace.
NB: This is done because of the way slider handles the popup (indicator).
It is not getting trapped in the evas functions (or rather, the popup
falls through the traps there) so framespace is not accounted on the
popup, resulting in improper popup position when run in wayland.

NB: This is the only widget known to need this.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@75546 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33