From: Kim Shinwoo <kimcinoo.efl@gmail.com> Subject: [E-devel] [patch][elementary] diskselector - select option, disabling auto selection. datetime - show even number of items
by default, an diskselector item which is right in the center of
diskselector is selected automatically.
and if the item style emit elm,action,click signal, the item is selected
also.

so there would be a case that item select callback is called by two
ways.
one is clicking the item,
the other is locating item right in the center of diskselector (if
number
of displayed item is odd, then after scrolling the selected callback is
called always).

because of this reason, datetime module always set number of displayed
item
to even number.

to resolve this issue, i add a api to enable or disable default
behavior -
the auto selection - that can be found on the attachment.
moreover the second attachment is for datetime module, now the
datetime can
show even number of items.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@75820 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33