From: Kim Shinwoo <kimcinoo.efl@gmail.com> Subject: [E-devel] [patch][elementary] index - add access feature
the attachment has access features for the index. and this patch is
depends
on the previous patch which has
_elm_access_edje_object_part_object_unregister();
to find the previous patch, please refer to the following

http://sourceforge.net/mailarchive/forum.php?thread_name=CAP-c0nG8NkMx3J-YkSJnWdB23cWOPrtk1Rx10Lae0Zwxst1eeQ%40mail.gmail.com&forum_name=enlightenment-devel

the index could have two levels, one is level 0, the other is level 1.
the
patch supports only level 0 because we have no style for the level 1..
(just a cowardly excuse, i will keep up with this.. and update..)



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@76532 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33